Remove shell for monkeysphere user (Closes: #901489)
As of 0.43, Monkeysphere will use 'runuser' instead of 'su' to perform
operation using the monkeysphere user. 'runuser' works when there is
no shell for the user.

When freshly installing, create a monkeysphere user without a
shell.  If the monkeysphere user is already present on the system the
shell for that user should be removed.

once debian buster is released, we should no longer need to change the
shell for the monkeysphere user.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>